Create FUNDING.yml
Vaccination for Covid-19 is a research project which lay emphasis on a leave extract (Annonia Muricata) which can be further refine to capsules and xyrub. The biochemical agent have been used to cure different aliments of diseases such as cancer, diabetes, fever, sore throat etc. The evidence have  been deduce and proved that it has the potential of  curing all the symptoms of Covid-19. This  research is conducted to meet the  aspiration of the health sector and to make a world a better place for the generation to come.